Last Commit before Refactor
Added Navbar, changed sizing for text, fix animations, got a new layout for the main part of the page, and more. I'm going to rewrite all of this tonight so I have some better sanity to not have everything break. I feel like if I was to do that it might work :/