Had to switch how the "current-container" was
Changed since the current-container needed to be on the bottom, taking the width of the page, this I feel like will be temporary for a little bit. I plan to style the grid a bit more but I have no idea what else I could add. Animations will be next for this section, but I got to get the other pages written and semi-styled for that.